Project was finalized and publicaded. Thank you for everyone who's helped me! <3

+++ src/app/pages/showed/editor/editor.component.ts
@@ -2,9 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Projeto } from '../../core/projeto.models';
 import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
 import { ProjetoService } from '../../core/projeto.service';
-import { ConfirmationService } from 'primeng/api';
-import { MessageService } from 'primeng/api';
-import { PrimeNGConfig } from 'primeng/api';
+import { ConfirmationService, MessageService, PrimeNGConfig } from 'primeng/api';
 import { NgxSpinnerService } from 'ngx-spinner';
 
 @Component({
@@ -14,23 +12,22 @@ import { NgxSpinnerService } from 'ngx-spinner';
 })
 export class EditorComponent implements OnInit {
   projetoList: Projeto[] = [];
-  
 
-  constructor(public ref: DynamicDialogRef, 
+  constructor(
+    public ref: DynamicDialogRef, 
     public config: DynamicDialogConfig,
     private projetoService: ProjetoService,
     private confirmationService: ConfirmationService,
     private messageService: MessageService,
     private primengConfig: PrimeNGConfig,
     private spinner: NgxSpinnerService
-  ) 
-    {
+  ) {
     this.projetoList = this.config.data?.projetoList || [];
   }
 
-  ngOnInit(){
+  ngOnInit() {
     this.primengConfig.setTranslation({
-      accept: 'Sim',                                 //configuração de tradução do popup de confirmação presente nos botões e na hora de POST,GET ou DELETE no backend
+      accept: 'Sim',
       reject: 'Não'
     });
   }
@@ -41,25 +38,37 @@ export class EditorComponent implements OnInit {
 
   cancelar() {
     this.ref.close();
+    
   }
 
-  confirmDelete(event: Event, id: number) {
+  confirmDelete(event: Event, id: number | undefined) {
+    if (id === undefined) {
+      console.error('ID do projeto é indefinido.');
+      return;
+    }
+  
     this.confirmationService.confirm({
       target: event.target || undefined,
       message: 'Você tem certeza que deseja excluir este projeto?',
       icon: 'pi pi-exclamation-triangle',
       accept: () => {
+        this.spinner.show();
         this.projetoService.deleteProjeto(id).subscribe(() => {
           this.projetoList = this.projetoList.filter(projeto => projeto.id !== id);
-          this.messageService.add({ severity: 'success', summary: 'Sucesso', detail: `Projeto com id ${id} deletado.`, life: 3000 });
-          this.spinner.show();
-          setTimeout(()=>{               
+          this.messageService.add({ severity: 'success', summary: 'Sucesso', detail: 'Projeto deletado com sucesso', life: 3000 });
+          
+          setTimeout(() => {
             this.spinner.hide();
-          }, 1000)
-        }, (error) => {
+            this.salvar(); // Atualizar a lista de projetos
+          }, 2000);
+          
+        }, error => {
+          this.spinner.hide();
           console.error('Erro ao deletar o projeto:', error);
-          this.messageService.add({ severity: 'error', summary: 'Erro', detail: 'Erro ao deletar o projeto.', life: 3000 });
         });
+      },
+      reject: () => {
+        this.messageService.add({ severity: 'info', summary: 'Cancelado', detail: 'Nenhuma alteração foi concluída', life: 3000 });
       }
     });
   }
